removed Locker+ObjectiveC
